Move the property parsing to the base class so the Triton invoker can use as well.

+++ test/shared_test/test_ortops_azure.cc
@@ -37,7 +37,7 @@ TEST(AzureOps, OpenAIWhisper) {
   std::vector<std::string> expected_output{"Thank you for pressing the self-destruct button",
                                            "ship will self-destruct in three minutes"};
 
-  OutputValidator find_strings_in_output = [&expected_output](int output_idx, Ort::Value& actual, TestValue expected) {
+  OutputValidator find_strings_in_output = [&expected_output](size_t output_idx, Ort::Value& actual, TestValue expected) {
     std::vector<std::string> output_string;
     GetTensorMutableDataString(Ort::GetApi(), actual, output_string);
